adding other uc

+++ NutritionalAdvice.Domain/MealPlans/IMealPlanFactory.cs
@@ -8,6 +8,6 @@ namespace NutritionalAdvice.Domain.MealPlans
 {
     public interface IMealPlanFactory
     {
-        MealPlan Create(string name, string description, string goal, int dailyCalories, int dailyProtein, int dailyCarbohydrates, int dailyFats, Guid nutritionistId, Guid patientId);
+        MealPlan Create(string name, string description, string goal, int dailyCalories, double dailyProtein, double dailyCarbohydrates, double dailyFats, Guid nutritionistId, Guid patientId);
     }
 }
